printer-requirements.txt
# printer
click
python-escpos==3.0a3
requests
arrow
